fixes
* fixed startup secret (it was not wiping out backing store)
* fixed rehydration logic
* other minor updates
